django/requirements.txt
django
django-froala-editor
django-allauth
python-dateutil
arc-library
smarturls
Pillow


